anchor stale socket cleanup identities

+++ internal/supervisor/rename_noreplace_other.go
@@ -3,9 +3,55 @@
 package supervisor
 
 import (
+	"fmt"
 	"os"
 	"path/filepath"
 )
+
+type pathIdentity struct {
+	info os.FileInfo
+}
+
+func pathIdentityAt(dir *os.File, name string) (pathIdentity, error) {
+	info, err := os.Lstat(filepath.Join(dir.Name(), name))
+	if err != nil {
+		return pathIdentity{}, err
+	}
+	return pathIdentity{info: info}, nil
+}
+
+func pathIdentityForFile(file *os.File) (pathIdentity, error) {
+	info, err := file.Stat()
+	if err != nil {
+		return pathIdentity{}, err
+	}
+	return pathIdentity{info: info}, nil
+}
+
+func samePathIdentity(left, right pathIdentity) bool {
+	return os.SameFile(left.info, right.info) && left.info.Mode().Type() == right.info.Mode().Type()
+}
+
+func (identity pathIdentity) isSocket() bool {
+	return identity.info.Mode()&os.ModeSocket != 0
+}
+
+func renameNoReplaceEntryAt(
+	oldDir *os.File,
+	oldName string,
+	newDir *os.File,
+	newName string,
+	expected pathIdentity,
+) error {
+	current, err := pathIdentityAt(oldDir, oldName)
+	if err != nil {
+		return err
+	}
+	if !samePathIdentity(current, expected) {
+		return fmt.Errorf("filesystem entry changed before restoration")
+	}
+	return renameNoReplaceAt(oldDir, oldName, newDir, newName)
+}
 
 func renameNoReplaceAt(oldDir *os.File, oldName string, _ *os.File, newName string) error {
 	return unsupportedRenameNoReplaceAt(oldDir, oldName, newName)
@@ -15,15 +61,14 @@ func moveToQuarantineAt(oldDir *os.File, oldName string, _ *os.File, newName str
 	return unsupportedRenameNoReplaceAt(oldDir, oldName, newName)
 }
 
-func pathIsSocketAt(dir *os.File, name string) (bool, error) {
-	info, err := os.Lstat(filepath.Join(dir.Name(), name))
+func removeAt(dir *os.File, name string, expected pathIdentity) error {
+	current, err := pathIdentityAt(dir, name)
 	if err != nil {
-		return false, err
+		return err
 	}
-	return info.Mode()&os.ModeSocket != 0, nil
-}
-
-func removeAt(dir *os.File, name string) error {
+	if !samePathIdentity(current, expected) {
+		return fmt.Errorf("filesystem entry changed before removal")
+	}
 	return os.Remove(filepath.Join(dir.Name(), name))
 }
 
@@ -32,14 +77,37 @@ func createPrivateDirAt(dir *os.File, name string) (*os.File, error) {
 	if err := os.Mkdir(path, 0700); err != nil {
 		return nil, err
 	}
+	created, err := os.Lstat(path)
+	if err != nil {
+		return nil, err
+	}
 	privateDir, err := os.Open(path)
 	if err != nil {
-		_ = os.Remove(path)
+		current, statErr := os.Lstat(path)
+		if statErr == nil && os.SameFile(created, current) {
+			_ = os.Remove(path)
+		}
 		return nil, err
+	}
+	opened, err := privateDir.Stat()
+	if err != nil {
+		_ = privateDir.Close()
+		return nil, err
+	}
+	if !os.SameFile(created, opened) {
+		_ = privateDir.Close()
+		return nil, fmt.Errorf("quarantine directory changed between creation and open")
 	}
 	return privateDir, nil
 }
 
-func removeDirAt(dir *os.File, name string) error {
+func removeDirAt(dir *os.File, name string, expected pathIdentity) error {
+	current, err := pathIdentityAt(dir, name)
+	if err != nil {
+		return err
+	}
+	if !samePathIdentity(current, expected) {
+		return fmt.Errorf("quarantine directory changed before removal")
+	}
 	return os.Remove(filepath.Join(dir.Name(), name))
 }
